jesd_status: Refactor status and lane info views
This changes the way the jesd devices are displayed. Instead of
displaying all the devices in the same lane and status window each
device is displayed separatedly (although still using the same windows).
The user is presented with option keys to move around devices. With this,
will also be easier to introduce jesd204c devices with slighly different
lables that can be changed as we move between devices.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>