jesd_status: Don't always refresh lane and status windows
There's no need to refresh the complete windows in every iteration. It
only makes sense to do it when moving between devices since the status
labels might change (with 204c support) and the lane_info might not
exist. This change also improves the tool when running on a low speed
emulated terminal (eg: running it over serial console) since it won't
flicker that much..

Signed-off-by: Nuno Sá <nuno.sa@analog.com>