jesd_lane_status: Revise lane latency computation
This fixes a bug where lane_latency_multiframes wasn't converted to
octets, due to the missing F parameter.
The previous limits were kind of arbitrarily chosen.
We now find the lane with the smallest latency, and use this as the
reference. Errors (deterministic latency) are when the latency is larger
than F * K. Critical is now considered 50% of F * K. This allows for
enough margin. In case DL is not required the Elastic Buffer size could
be another measure. However in all our designs EBS is 256 which is
typically at least 2 time F * K.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>