cmake: make jesd_status and jesd_eye_scan selectable
Add options to make jesd_status and jesd_eye_scan selectable. This is
useful for systems (like meta-adi) where we are only interested in
jesd_status and so we do not want to build/install jesd_eye_scan (and
more importantly gtk+3).

Obviously, one of the binaries needs to be enabled.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>